Split decoders into own file
Still need to convert the decoders to PyTorch/make timeseries decoder
